+++ src/Infolists/Components/Fieldset.php
@@ -17,7 +17,8 @@ class Fieldset extends Component
      */
     protected array $schema = [];
 
-    protected int $columns = 12;
+    /** @var int|array<string, int> */
+    protected int|array $columns = 12;
 
     public function __construct(mixed $label)
     {
@@ -39,7 +40,8 @@ class Fieldset extends Component
         return $this;
     }
 
-    public function columns(int $columns): static
+    /** @param int|array<string, int> $columns */
+    public function columns(int|array $columns): static
     {
         $this->columns = $columns;