+++ src/Infolists/Components/Section.php
@@ -29,7 +29,8 @@ class Section extends Component
      */
     protected array $actions = [];
 
-    protected int $columns = 12;
+    /** @var int|array<string, int> */
+    protected int|array $columns = 12;
 
     public function __construct(mixed $heading)
     {
@@ -67,7 +68,8 @@ class Section extends Component
         return $this;
     }
 
-    public function columns(int $columns): static
+    /** @param int|array<string, int> $columns */
+    public function columns(int|array $columns): static
     {
         $this->columns = $columns;