+++ src/Infolists/Infolist.php
@@ -19,7 +19,8 @@ class Infolist implements Arrayable, JsonSerializable
 
     protected ?Model $record = null;
 
-    protected int $columns = 1;
+    /** @var int|array<string, int> */
+    protected int|array $columns = 1;
 
     protected string $operation = 'view';
 
@@ -77,7 +78,8 @@ class Infolist implements Arrayable, JsonSerializable
         return $this;
     }
 
-    public function columns(int $columns): static
+    /** @param int|array<string, int> $columns */
+    public function columns(int|array $columns): static
     {
         $this->columns = $columns;